refactor: rename standalone soc_agent_tier2 to a2a_tier2 and clean up redundant mapping CLI logic
